Add unit test for flux detection limit using N2O eddy covariance data
Introduced `test_flux_detection_limit` method in `test_createvar.py` to validate the flux detection workflow. Includes data preparation, covariance calculation, and validation of results such as lag range, detection limit, and noise RMSE.

+++ diive/pkgs/echires/fluxdetectionlimit.py
@@ -168,7 +168,7 @@ class FluxDetectionLimit:
             ts_col: str,
             h2o_col: str,
             press_col: str,
-            default_lag: int,
+            default_lag: float,
             noise_range: int,
             lag_range: list,
             lag_stepsize: int,
